+++ src/postgres/queries.rs
@@ -4,10 +4,11 @@ mod select;
 mod update;
 
 pub use delete_from::{delete_from, DeleteFrom};
-pub use insert_into::{insert_into, InsertInto, Values};
+pub use insert_into::{insert_into, BareInsertInto, InsertInto, Values};
 pub use select::{select, Joinable, Orderable, Select};
-pub use update::{update, Update, BareUpdate};
+pub use update::{update, BareUpdate, Update};
 
 pub(crate) use delete_from::delete_from_with;
+pub(crate) use insert_into::insert_into_with;
 pub(crate) use select::select_with;
 pub(crate) use update::update_with;